Fold model_freshness into SystemService
A standalone freshness_service.py for one method + one dataclass was too
narrow. model_freshness is the same conceptual family as data inventory
and last_import_at — it belongs on the existing SystemService alongside
status(), which already aggregates pipeline-state queries.
